pyproject.toml
[project]
name = "insideLLMs"
version = "0.2.0"
description = "Cross-model behavioural probe harness for LLM evaluation and comparison."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "insideLLMs Contributors" }
]
keywords = [
    "llm",
    "language-models",
    "nlp",
    "machine-learning",
    "ai",
    "evaluation",
    "benchmarking",
    "testing",
    "probing",
    "harness",
    "comparison",
    "behavioural",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
    "Typing :: Typed",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "transformers>=4.30.0",
    "huggingface-hub>=0.16.0",
    "pyyaml>=6.0",
    "tuf>=3.0.0",
    "oras>=0.2.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
nlp = [
    "nltk>=3.8.0",
    "spacy>=3.5.0",
    "scikit-learn>=1.2.0",
    "gensim>=4.3.0",
]
visualization = [
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "seaborn>=0.12.0",
]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
    "types-PyYAML>=6.0.12.20240311",
]
langchain = [
    "langchain",
    "langchain-core",
    "langgraph",
]
serving = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]
all = [
    "insideLLMs[nlp,visualization,dev,serving]",
]

[project.scripts]
insidellms = "insideLLMs.cli:main"

[project.urls]
Homepage = "https://github.com/dr-gareth-roberts/insideLLMs"
Repository = "https://github.com/dr-gareth-roberts/insideLLMs"
Documentation = "https://dr-gareth-roberts.github.io/insideLLMs/"
"Changelog" = "https://github.com/dr-gareth-roberts/insideLLMs/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/dr-gareth-roberts/insideLLMs/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["insideLLMs*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    ".backups",
    ".tmp",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function calls in default arguments
    "ARG001", # unused function argument
    "ARG002", # unused method argument
    "UP035",  # prefer collections.abc import (not enforced for now)
    "UP045",  # Optional[T] -> T | None (not enforced for now)
]

[tool.ruff.lint.isort]
known-first-party = ["insideLLMs"]

[tool.ruff.lint.per-file-ignores]
"examples/**/*.py" = ["ARG", "F401", "UP"]
"tests/**/*.py" = ["ARG", "B017", "E741", "F401", "SIM", "UP"]
"insideLLMs/caching.py" = ["E402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
ignore_missing_imports = true
pretty = true
show_error_codes = true
show_column_numbers = true
disable_error_code = [
    "attr-defined",
    "arg-type",
    "assignment",
    "call-arg",
    "call-overload",
    "dict-item",
    "exit-return",
    "has-type",
    "list-item",
    "name-defined",
    "operator",
    "index",
    "var-annotated",
    "no-any-return",
    "misc",
    "union-attr",
    "return-value",
    "override",
    "syntax",
    "type-var",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["insideLLMs"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
